Add files from different laptop
Have worked on a different machine, now adding all files to this repo so I can continue on different machine. Set up the index.html with the necessary content and have started with style.css (fonts, some of the positioning, background and font colors), added a few images to give the file a realistic feel.