Make the layout responsive and work well
I removed the flex: 1 from all elements in hero and header, and made the
elements not dependent on screen height but rather screen width.
This is better because you would make the window less wide and not less
tall. Furthermore, I turned on wrap for the hero container, for more
responsiveness. I stepped away from giving the elements a width, and
kept it only for the image.
The hero is now centered, so it wraps correctly, this is a stylistic
choice that can be avoided with media queries based on the screen width.