Format the quote section and make it responsive
Added margin, a max-width, and column direction to a new div that
contains both the quote main text as well as the quote author. This
allows me to align the quote author where it belongs according to the
spec. Is also responsive. Not sure if I need the "quote" div?