Format the CTA and Footer section
Made the CTA a flex container with a max-height, so that the button gets
pushed to the right of the text (no space available). I also made it
wrap so it is reponsive. Used "> *" syntax to remove the margin and
padding for all child elements of ctatext (headline and copy).
Footer was centered and got some padding. Now needs double checking of
the entire site at various screen sizes, then host on Github.